ReferenceApplication Consumer Architecture must be improved #282 - Encoding moved to separate project - Changed the Consumer architecture - it is ready to be moved to separate project - There are composition errors.
