--- Networking/Networking.SimulatorInteroperabilityTest/Settings.cs
@@ -1,8 +1,0 @@
-﻿
-namespace UAOOI.Networking.SimulatorInteroperabilityTest
-{
-  internal static class Settings
-  {
-    internal const string ProducerConfigurationRepositoryGroup = "repositoryGroup";
-  }
-}
