+++ Networking/ReferenceApplication/AppBootstrapper.cs
@@ -6,7 +6,6 @@ using System.ComponentModel.Composition.Hosting;
 using System.Diagnostics;
 using System.Reactive.Disposables;
 using System.Windows;
-using UAOOI.Networking.ReferenceApplication.Consumer;
 using UAOOI.Networking.ReferenceApplication.Diagnostic;
 using UAOOI.Networking.ReferenceApplication.MEF;
 using UAOOI.Networking.ReferenceApplication.Producer;
@@ -58,10 +57,10 @@ namespace UAOOI.Networking.ReferenceApplication
         m_EventSourceBootstrapper = Container.GetExportedValue<EventSourceBootstrapper>();
         m_EventSourceBootstrapper.Run(EventEntry => m_Log.Add(EventEntry));
         ReferenceApplicationEventSource.Log.StartingApplication(Settings.Default.MessageHandlerProvider);
-        ConsumerDataManagementSetup m_ConsumerConfigurationFactory = Container.GetExportedValue<ConsumerDataManagementSetup>();
+        Consumer.ConsumerDataManagementSetup m_ConsumerConfigurationFactory = Container.GetExportedValue<Consumer.ConsumerDataManagementSetup>();
         m_ConsumerConfigurationFactory.Setup();
         m_Components.Add(m_ConsumerConfigurationFactory);
-        ReferenceApplicationEventSource.Log.PartCreated(nameof(ConsumerDataManagementSetup));
+        ReferenceApplicationEventSource.Log.PartCreated(nameof(Consumer.ConsumerDataManagementSetup));
         OPCUAServerProducerSimulator m_OPCUAServerProducerSimulator = Container.GetExportedValue<OPCUAServerProducerSimulator>();
         m_OPCUAServerProducerSimulator.Setup();
         m_Components.Add(m_OPCUAServerProducerSimulator);
