+++ Networking/ReferenceApplication/Consumer/ConsumerCompositionSettings.cs
@@ -1,0 +1,13 @@
+﻿namespace UAOOI.Networking.ReferenceApplication.Consumer
+{
+  internal class ConsumerCompositionSettings
+  {
+
+    internal const string ConfigurationFileNameContract = "ConsumerCompositionSettings.ConfigurationFileName";
+    internal const string ConfigurationFactoryContract = "ConsumerCompositionSettings.ConfigurationFactory";
+    internal const string EncodingFactoryContract = "ConsumerCompositionSettings.EncodingFactory";
+    internal const string BindingFactoryContract = "ConsumerCompositionSettings.BindingFactory";
+    internal const string ViewModelContract = "ConsumerCompositionSettings.ConsumerViewModel";
+
+  }
+}
