+++ Networking/ReferenceApplication/MEF/DefaultServiceRegistrar.cs
@@ -26,7 +26,8 @@ namespace UAOOI.Networking.ReferenceApplication.MEF
     {
       return new List<ComposablePartCatalog>(new ComposablePartCatalog[] { new AssemblyCatalog(Assembly.GetAssembly(typeof(AppBootstrapper))),
                                                                            new AssemblyCatalog(Assembly.GetAssembly(typeof(DataManagementSetup))),
-                                                                           new AssemblyCatalog(Assembly.GetAssembly(typeof(UAOOI.Networking.SimulatorInteroperabilityTest.EncodingFactoryBinarySimple)))});
+                                                                           new AssemblyCatalog(Assembly.GetAssembly(typeof(UAOOI.Networking.Encoding.CompositionSettings)),
+                                                                           new AssemblyCatalog(Assembly.GetAssembly(typeof(UAOOI.Networking.SimulatorInteroperabilityTest.ProducerCompositionSettings))))});
     }
 
   }
