+++ Networking/ReferenceApplication/MainWindow.xaml
@@ -109,12 +109,12 @@
                     <Label Style="{StaticResource TextBox}" >Consumer status</Label>
                 </StackPanel>
                 <StackPanel Orientation="Vertical"  Grid.Column="1" Grid.Row="0">
-                    <TextBox Style="{StaticResource TextBox}" Text="{Binding Path=ConsumerFramesReceived, FallbackValue=?????, Mode=OneWay}" ToolTip="Packages received by the consumer from all sources"/>
-                    <TextBox Style="{StaticResource TextBox}" Text="{Binding Path=ConsumerReceivedBytes, FallbackValue=?????, Mode=OneWay }" ToolTip="Bytes received by the consumer from all sources"/>
-                    <TextBox  Style="{StaticResource TextBox}" Text="{Binding Path=ConsumerErrorMessage, Mode=TwoWay }" ToolTip="Bytes sent by the producer"/>
-                    <Button  Style="{StaticResource TextBox}" Command="{Binding Path=ConsumerUpdateConfiguration}" >Update Configuration</Button>
+                    <TextBox Style="{StaticResource TextBox}" Text="{Binding Path=ConsumerViewModel.ConsumerFramesReceived, FallbackValue=?????, Mode=OneWay}" ToolTip="Packages received by the consumer from all sources"/>
+                    <TextBox Style="{StaticResource TextBox}" Text="{Binding Path=ConsumerViewModel.ConsumerReceivedBytes, FallbackValue=?????, Mode=OneWay }" ToolTip="Bytes received by the consumer from all sources"/>
+                    <TextBox  Style="{StaticResource TextBox}" Text="{Binding Path=ConsumerViewModel.ConsumerErrorMessage, Mode=TwoWay }" ToolTip="Bytes sent by the producer"/>
+                    <Button  Style="{StaticResource TextBox}" Command="{Binding Path=ConsumerViewModel.ConsumerUpdateConfiguration}" >Update Configuration</Button>
                 </StackPanel>
-                <ListBox Grid.Column="0" Grid.Row="1" Grid.ColumnSpan="2" ItemsSource="{Binding Path=ConsumerLog}" VerticalAlignment="Stretch" VerticalContentAlignment="Stretch"  />
+                <ListBox Grid.Column="0" Grid.Row="1" Grid.ColumnSpan="2" ItemsSource="{Binding Path=ConsumerViewModel.ConsumerLog}" VerticalAlignment="Stretch" VerticalContentAlignment="Stretch"  />
             </Grid>
         </Grid>
         <StackPanel Orientation="Vertical" Grid.Row="2" Background="#FF0AC3F9" >
