+++ Networking/ReferenceApplication/Producer/ProducerCompositionSettings.cs
@@ -8,7 +8,7 @@ namespace UAOOI.Networking.ReferenceApplication.Producer
     public const string ConfigurationFactoryContract = "ProducerCompositionSettings.ConfigurationFactory";
     public const string EncodingFactoryContract = "ProducerCompositionSettings.EncodingFactory";
     public const string BindingFactoryContract = "ProducerCompositionSettings.BindingFactory";
-    public const string ProducerViewModelContract = "ProducerCompositionSettings.ProducerViewModel";
+    public const string ViewModelContract = "ProducerCompositionSettings.ProducerViewModel";
 
   }
