+++ Networking/ReferenceApplication/Producer/ProducerSettings.cs
@@ -5,11 +5,13 @@ namespace UAOOI.Networking.ReferenceApplication.Producer
 {
 
   [Export]
-  [PartCreationPolicy(CreationPolicy.NonShared)]
+  [PartCreationPolicy(CreationPolicy.Shared)]
   public class ProducerSettings
   {
+
     [Export(ProducerCompositionSettings.ConfigurationFileNameContract)]
     public string ProducerConfigurationFileName => Properties.Settings.Default.ProducerConfigurationFileName;
+
   }
 
 }
