+++ Networking/ReferenceApplication/README.MD
@@ -27,7 +27,7 @@ Detailed description is available in file [ReadMe.txt](ReadMe.txt)
 
 ### Troubleshooting
 
-For the diagnostic purpose the application supports network tracing that provides access to information about method invocations and network traffic generated by the application. This feature is useful for debugging applications under development as well as for analyzing deployed applications. The output provided by network tracing is customizable to support different usage scenarios at development time and in a production environment. Traces are to publish specific messages that are generated during application execution. To learn more about configuration of the tracing visit the documentation at [Network Tracing in the .NET Framework](https://msdn.microsoft.com/en-us/library/hyb3xww8(v=vs.110).aspx). 
+For the diagnostic purpose the application supports network tracing that provides access to information about method invocations and network traffic generated by the application. This feature is useful for debugging applications under development as well as for analyzing deployed applications. The output provided by network tracing is customizable to support different usage scenarios at development time and in a production environment. Traces are to publish specific messages that are generated during application execution. To learn more about configuration of the tracing visit the documentation at [Network Tracing in the .NET Framework](https://msdn.microsoft.com/en-us/library/hyb3xww8.aspx). 
 
 The log file `UANetworkingReferenceApplication.log` captures output messages from some public methods of the `Socket`, `TcpListener`, `TcpClient`, and `Dns` classes. To find the log file location click the `Actions.Open Configuration folder` menu entry.
