+++ UAObjectOrientedInternet.sln
@@ -81,7 +81,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Networking.SimulatorInterop
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SemanticData.InformationModelFactory", "SemanticData\InformationModelFactory\SemanticData.InformationModelFactory.csproj", "{6D20785D-0451-46A9-9AC2-EC66B7C0D18A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SemanticData.UANodeSetValidation", "SemanticData\UANodeSetValidation\SemanticData.UANodeSetValidation.csproj", "{352378A0-8B2E-4679-BB44-FD0BCC8467D9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SemanticData.UANodeSetValidation", "SemanticData\UANodeSetValidation\SemanticData.UANodeSetValidation.csproj", "{352378A0-8B2E-4679-BB44-FD0BCC8467D9}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SemanticData.AddressSpaceComplianceTestTool", "SemanticData\AddressSpaceComplianceTestTool\SemanticData.AddressSpaceComplianceTestTool.csproj", "{05DB9E60-B12B-4AA3-8A7C-514F87CFF464}"
 EndProject
@@ -90,6 +90,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SemanticData.DataSerialization.UnitTest", "SemanticData\Tests\DataSerializationUnitTestProject\SemanticData.DataSerialization.UnitTest.csproj", "{268DEFE0-D604-4888-BF69-57EBAE0FD7E7}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SemanticData.UANodeSetValidation.UnitTest", "SemanticData\Tests\USNodeSetValidationUnitTestProject\SemanticData.UANodeSetValidation.UnitTest.csproj", "{2DC898B5-AFD9-47B2-A962-C4002C1AE8FF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Networking.Encoding", "Networking\Networking.Encoding\Networking.Encoding.csproj", "{12F06CE3-1BFB-49FC-AA15-AF3A1FF1C66A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -197,6 +199,10 @@ Global
 		{2DC898B5-AFD9-47B2-A962-C4002C1AE8FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2DC898B5-AFD9-47B2-A962-C4002C1AE8FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2DC898B5-AFD9-47B2-A962-C4002C1AE8FF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{12F06CE3-1BFB-49FC-AA15-AF3A1FF1C66A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{12F06CE3-1BFB-49FC-AA15-AF3A1FF1C66A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{12F06CE3-1BFB-49FC-AA15-AF3A1FF1C66A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{12F06CE3-1BFB-49FC-AA15-AF3A1FF1C66A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -229,6 +235,7 @@ Global
 		{37D26B63-74FE-45C6-9F61-251BDF69EDF7} = {AF01F501-9B19-4357-B8A7-4EFC782BF1AF}
 		{268DEFE0-D604-4888-BF69-57EBAE0FD7E7} = {AF01F501-9B19-4357-B8A7-4EFC782BF1AF}
 		{2DC898B5-AFD9-47B2-A962-C4002C1AE8FF} = {AF01F501-9B19-4357-B8A7-4EFC782BF1AF}
+		{12F06CE3-1BFB-49FC-AA15-AF3A1FF1C66A} = {F4178F51-3BFF-46DE-841D-C01D4ABA9B0C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9388D93E-9AA9-40EE-A0C6-F096265079B2}
